formatting, render() changes (indentation level of tags)
I changed some of the formatting for myself so that I'd personally read
it better. If you like the changes you can accept it.

I merged buildcontent() with render() for conciseness, seeing that
building content simply renders children.

I implemented indentation level for tags, so that the source code in
html will show up as properly formatted.